requirements.txt
discord.py
python-dotenv
fuzzywuzzy
requests
beautifulsoup4